requirements.txt
inquirer
openai
python-dotenv